style : #32 Calendar Header 스타일 적용
